endret leggTilPrøve og redigePrøve slik at de bruker sql endret fra å bruke Float ti Dpuble som tallformat

+++ Retteark/Klasser og strukturer/Karaktergrense.swift
@@ -11,10 +11,10 @@ import Foundation
 class Karaktergrense: Hashable, Codable, Identifiable {
     var id = UUID().uuidString
     var karakter: String
-    var grense: Float?
+    var grense: Double?
     
     
-    init(karakter: String, grense: Float) {
+    init(karakter: String, grense: Double) {
         self.karakter = karakter
         self.grense = grense
     }
